:star: Add define mission geomtry
On the mission plan you can now define a geometry for the area or axis
you want to chart with the RPA. This geometry is drawn in the map and
then saved in as a corresponding geoJSON feature with all it's data,
this gemoetry is also editable and updated in the database accordingly.
The map also detects if there is an existing geometry in the database
and it draws it and makes it editable for the user. The map also detects
the type of mission to show the appropiate controls and disallows the
creation of more than one feature on one mission at the same time.