Make ground elevation calculations ignore landing paths
For the waypoints this utility calculates the ground elevation for this
waypoint and calculates the height relative to take off as the desired
height plus the diference in ground elevations between this WP and the
take of point.

For the landing path we want the `altRelative` to be fixed to protect
the decending slope calculated.
So in case this function is called again, for the landing path WP it
calculates the Absolute Height as the `altRelative` minus the difference
in ground elevations between the take off point and this WP. This way we
get the height the RPA is going to have at this WP, and if `altAbsolute`
goes below zero it means impact has occured.